v1

+++ scripts/train_vqgan.py
@@ -13,10 +13,27 @@ def main():
     parser = argparse.ArgumentParser()
     parser = pl.Trainer.add_argparse_args(parser)
     parser = VQGAN.add_model_specific_args(parser)
+
+    parser = argparse.ArgumentParser(parents=[parser], add_help=False)
+    parser.add_argument('--data_path', type=str, default='/datasets01/Kinetics400_Frames/videos')
+    parser.add_argument('--sequence_length', type=int, default=16)
+    parser.add_argument('--resolution', type=int, default=64)
+    parser.add_argument('--batch_size', type=int, default=32)
+    parser.add_argument('--num_workers', type=int, default=8)
+    parser.add_argument('--image_channels', type=int, default=3)
+    parser.add_argument('--smap_cond', type=int, default=0)
+    parser.add_argument('--smap_only', action='store_true')
+    parser.add_argument('--text_cond', action='store_true')
+    parser.add_argument('--vtokens', action='store_true')
+    parser.add_argument('--vtokens_pos', action='store_true')
+    parser.add_argument('--spatial_length', type=int, default=15)
+    parser.add_argument('--sample_every_n_frames', type=int, default=1)
+    parser.add_argument('--image_folder', action='store_true')
+    parser.add_argument('--stft_data', action='store_true')
     # parser = VideoData.add_data_specific_args(parser)
-    args = parser.parse_args('/root/autodl-tmp/')
 
     # data = VideoData(args)
+    args = parser.parse_args()
 
     train_data_config = {
         'target': 'tats.VideoDataset',
@@ -25,8 +42,8 @@ def main():
                 'folder': '/root/autodl-tmp/train_split_frame',
                 'image_width': 480,
                 'image_height': 272,
-                'cube_size': 160,
-                'clip_size': 4,
+                'cube_size': args.resolution,
+                'clip_size': args.sequence_length,
             }
     }
     val_data_config = {
@@ -36,18 +53,18 @@ def main():
                 'folder': '/root/autodl-tmp/test_split_frame',
                 'image_width': 480,
                 'image_height': 272,
-                'cube_size': 160,
-                'clip_size': 4,
+                'cube_size': args.resolution,
+                'clip_size': args.sequence_length,
             }
     }
     
 
 
     data = DataModuleFromConfig(
-        batch_size = 4,
+        batch_size = args.batch_size,
         train = train_data_config,
         validation = val_data_config,
-        num_workers = 16,
+        num_workers = args.num_workers,
     )
 
     # pre-make relevant cached files if necessary
@@ -68,7 +85,7 @@ def main():
     callbacks.append(ModelCheckpoint(monitor='val/recon_loss', save_top_k=3, mode='min', filename='latest_checkpoint'))
     callbacks.append(ModelCheckpoint(every_n_train_steps=3000, save_top_k=-1, filename='{epoch}-{step}-{train/recon_loss:.2f}'))
     callbacks.append(ModelCheckpoint(every_n_train_steps=10000, save_top_k=-1, filename='{epoch}-{step}-10000-{train/recon_loss:.2f}'))
-    callbacks.append(ImageLogger(batch_frequency=750, max_images=4, clamp=True))
+    # callbacks.append(ImageLogger(batch_frequency=750, max_images=4, clamp=True))
     callbacks.append(VideoLogger(batch_frequency=1500, max_videos=4, clamp=True))
 
     kwargs = dict()
@@ -87,16 +104,17 @@ def main():
                 log_folder = folder
         if len(log_folder) > 0:
             ckpt_folder = os.path.join(base_dir, log_folder, 'checkpoints')
-            for fn in os.listdir(ckpt_folder):
-                if fn == 'latest_checkpoint.ckpt':
-                    ckpt_file = 'latest_checkpoint_prev.ckpt'
-                    os.rename(os.path.join(ckpt_folder, fn), os.path.join(ckpt_folder, ckpt_file))
-            if len(ckpt_file) > 0:
-                args.resume_from_checkpoint = os.path.join(ckpt_folder, ckpt_file)
-                print('will start from the recent ckpt %s'%args.resume_from_checkpoint)
+            if os.path.exists(ckpt_folder):
+                for fn in os.listdir(ckpt_folder):
+                    if fn == 'latest_checkpoint.ckpt':
+                        ckpt_file = 'latest_checkpoint_prev.ckpt'
+                        os.rename(os.path.join(ckpt_folder, fn), os.path.join(ckpt_folder, ckpt_file))
+                if len(ckpt_file) > 0:
+                    args.resume_from_checkpoint = os.path.join(ckpt_folder, ckpt_file)
+                    print('will start from the recent ckpt %s'%args.resume_from_checkpoint)
 
-    trainer = pl.Trainer.from_argparse_args(args, callbacks=callbacks, 
-                                            max_steps=args.max_steps, **kwargs)
+    trainer = pl.Trainer.from_argparse_args(args, callbacks=callbacks,  strategy = "ddp",
+                                            max_steps=args.max_steps, precision=16, amp_backend='native', **kwargs)
 
     trainer.fit(model, data)
